autoparts_backend/requirements.txt
absl-py==1.2.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiormq==6.4.2
aiosignal==1.3.1
anyio==3.5.0
APScheduler==3.9.1.post1
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
awsebcli==3.20.3
Babel==2.9.1
backcall==0.2.0
bidict==0.22.1
boto3==1.26.114
botocore==1.29.114
CacheControl==0.12.11
cachetools==5.2.0
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
ChatterBot==1.0.2
chatterbot-corpus==1.2.0
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==1.9.2
cryptography==40.0.2
cycler==0.11.0
dask==2022.10.2
decorator==5.1.1
dj-database-url==1.0.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==1.16.0
docopt==0.6.2
executing==0.9.1
fastapi==0.75.0
fbmessenger==6.0.0
fire==0.5.0
Flask==2.0.3
flatbuffers==23.3.3
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2023.4.0
future==0.16.0
gast==0.4.0
geographiclib==2.0
geopy==2.3.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.3
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==0.10.0
joblib==1.1.0
jsonpickle==3.0.1
jsonschema==4.17.3
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
locket==1.0.0
magic-filter==1.0.9
Markdown==3.4.1
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.3
mattermostwrapper==2.2
msgpack==1.0.5
multidict==5.2.0
networkx==2.6.3
nltk==3.8.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==1.4.1
parso==0.8.3
partd==1.4.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
Pint==0.20.1
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.9.0
pydot==1.4.2
Pygments==2.12.0
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pypiwin32==223
pyreadline3==3.4.1
pyrsistent==0.19.3
pytesseract==0.3.8
python-crfsuite==0.9.9
python-dateutil==2.7.5
python-decouple==3.6
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa-sdk==3.5.1
redis==4.5.4
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.0.2
scipy==1.8.0
semantic-version==2.8.5
sentry-sdk==1.14.0
six==1.14.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.1
sniffio==1.2.0
SQLAlchemy==1.2.19
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.17.1
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.65.0
traitlets==5.3.0
twilio==7.14.2
typeguard==3.0.2
typing-utils==0.1.0
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.11
uvicorn==0.17.6
wcwidth==0.1.9
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.0.3
whitenoise==6.2.0
wrapt==1.14.1
yarl==1.8.2
